shared/website-bot/requirements.txt
locust>=2.35.0
locust-plugins[playwright]==5.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-logging==0.42b0
requests==2.31.0
protobuf==4.24.4
grpcio==1.59.2
